Rename page file
