+++ client/main.jsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router";
-import { AppPage } from "./pages/app";
+import { AppPage } from "./pages/AppPage";
 import "./base.css";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
